+++ internal/provider/types/service_publisher_assignment_item.go
@@ -6,8 +6,8 @@ import "github.com/hashicorp/terraform-plugin-framework/types"
 
 type ServicePublisherAssignmentItem struct {
 	Primary             types.Bool                                 `tfsdk:"primary"`
-	PublisherExternalID types.Int64                                `tfsdk:"publisher_external_id"`
+	PublisherExternalID types.Int32                                `tfsdk:"publisher_external_id"`
 	PublisherName       types.String                               `tfsdk:"publisher_name"`
 	Reachability        *PublishersPrivateAppsResponseReachability `tfsdk:"reachability"`
-	ServiceExternalID   types.Int64                                `tfsdk:"service_external_id"`
+	ServiceExternalID   types.Int32                                `tfsdk:"service_external_id"`
 }
